Usefull tools for network admins
Some usefull tools for network admins.

doBackupFortigate - Script that connect to a fortigate appliance and do
a backup. The script also do a rotate of old backup files.

doBackupJuniper - Script that connect to a juniper router and do a
backup. The script also do a rotate of old backup files.